IP para ring configurable
Se puede enviar como parámetro una dirección IP para alcanzar al servidor siguiente en el anillo.
A falta de este, se seguirá tratando localmente. Servidores ahora comparten su dirección IP (la
cual averiguan mediante el nuevo obtener_ip_propia) cuando esta se encuentra configurada y requieren
que alguien se conecte a ellos.

- Verificación de input de sid y sid_sig. MAX_SID (actualmente 500)
- Emprolijo ringmsg_t; opcional priorización de mensajes (poco uso actual)
- Actualmente no hay reintento de conexión caída; simplemente se la reemplaza cerrando el ring gap
- Mejores mensajes
- Limpieza variada
- Unifico documentación interna de v1

Los archivos de documentación no deberían tratarse como finales ni actualizados, ya que son usados
como referencia interna ocasional, y pueden no estar al día o incluso contradecir con lo que hace el
código. Aprovechar los comentarios en el código fuente y basarse solo como guía en la doc no
oficial.

+++ common/ipc/shm.c
@@ -5,7 +5,7 @@
 #include "resources.h"
 #include "../log/log.h"
 
-int creashm(int key, int size) {
+int creashm(int key, size_t size) {
     key_t clave;
     clave = ftok(IPC_DIRECTORY, key);
     if (clave < 0) {
