+++ common/ipc/shm.h
@@ -1,7 +1,7 @@
 #ifndef BROKER_SHM_H
 #define BROKER_SHM_H
 
-int   creashm(int key, int size);
+int   creashm(int key, size_t size);
 int   getshm(int id);
 void* mapshm(int id);
 int   unmapshm(void* addr);
